Ignore unknown pkt types, instead of erroring out.
This change modifies the behaviour of the driver in case an
unknown packet type is encountered. Previously, the deserialisation
routine would raise an exception, which would cause the driver
to print an error and exit.

In the current implementation the unknown packet type(s) are
stored, their payload ignored, and the driver reports the fact that
unknown type(s) was/were received (with a request to report the
warning, ideally to the package maintainer).

This is not a solution for #16, although that issue prompted this
change.