driver: remove Beautifulsoup dependency. Fix #36.
This change reimplements the load_joint_offsets(..) function using
lxml. A simple XPath expression is used to access the calibration_offset
elements.

Run-time behaviour should be identical. Tested with ur-sim.

Note: I opted to use lxml instead of urdfdom_py, as the latter has
some issues parsing the UR5/10 urdfs and we only need access to a single
custom element. Rosdep keys for python-lxml were recently added (for
urdfdom) for all (Ubuntu) versions, so this should not be an issue.