description: add ROS-I base and tool0 frames. Fix #49 and #95.
Note that 'base' is essentially 'base_link' but rotated by 180
degrees over the Z-axis. This is necessary as the visual and
collision geometries appear to also have their origins rotated
180 degrees wrt the real robot.

'tool0' is similar to 'ee_link', but with its orientation such
that it coincides with an all-zeros TCP setting on the UR
controller. Users are expected to attach their own TCP frames
to this frame, instead of updating it (see also [1]).

[1] http://wiki.ros.org/Industrial/Tutorials/WorkingWithRosIndustrialRobotSupportPackages#Standardised_links_.2BAC8_frames