Added #define-based namespaces to prevent linking errors.
In the current setup, it is impossible to link against the ur5 or ur10 versions of the kinematics library, as all expose the same ur_kinematics::inverse() signature.
Since all libraries are exported in the catkin_package() macro, this results in undefined behavior when linking.
GCC just picks the first suitable signature it encounters, in this case the ur3 version, and continues without an error.
Adding a namespace allows the linker to find the correct library call.