gazebo: refactor gazebo and gazebo_ros_control integration.
This changes all bringup launch files in this package to be like those of a regular driver. In essence, Gazebo (and the UR(s) simulated by it) are treated as just-another-robot, with the associated launch file infrastructure.

As ur_robot_driver is currently the driver most users will be exposed to and have experience with, the launch file structure of that driver has been used as a template. But instead of starting the actual ur_robot_driver node, the 'lowest' launch file (ie: 'ur_control.launch') starts Gazebo and loads and spawns the controllers. After this launch file has completed its work, users should see a ROS API (ie: topics, services and actions) similar to what they'd see when using the a real UR with ur_robot_driver.

Main differences would be (at this point, this may change in the future):

 - no force-torque sensor output (ie: no 'wrench' topic)
 - no URScript topic (as the controller is not simulated by Gazebo)
 - no IO topics nor services
 - no MainBoard or similar topics
 - no Dashboard services
 - no 'tool communication' topics nor services
 - no 'tool0_controller' TF frame

Parameter files used with the xacro macro for the real robot (ie: joint limits, kinematics, visual and physical parameters) can be used with the Gazebo simulation as well.