gazebo: load URDF in bringup launch files.
The main 'driver' launch file should not load the URDF, as it's not the task of the driver.

This also removes one level of argument-forwarding (namely: the arguments for the xacro macro).

Finally, it provides a nicer template for users to base their own launch files on, as they 'only' have to deal with these top-level bringup launch files.