gazebo: restructure bringup entry points.
By lifting some of the nodes and parameters to the bringup files (from ur_common.launch.xml), they should be easier to use (and maintain).

Anything specific to the use of a robot in a particular configuration (ie: a workcell or simulation) should exist at the bringup level, not in a common (ie: shared) file, as that will make it harder to change (as changes will be shared by many other setups).

This restructuring makes ur_common.launch.xml redundant, so we also remove it.