Add tool0_controller frame for gazebo simulation
In order to make the simulation more consistent to the ur_robot_driver
(and ur_modern_driver) this adds a tool0_controller frame.
For the driver this corresponds to the tool transformation published by
the robot controller directly. As this doesn't exist for the simulated
robot, this adds the frame using an identity transform from tool0.